% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfwlinks-package.R
\docType{package}
\name{gfwlinks-package}
\alias{gfwlinks}
\alias{gfwlinks-package}
\title{gfwlinks: Build Global Fishing Watch Map URLs}
\description{
Build shareable Global Fishing Watch map links from vessel ids, without
opening the map and clicking through the UI. See \code{\link[=vessel_profile_url]{vessel_profile_url()}}
and \code{\link[=vessel_map_url]{vessel_map_url()}}.
}
\author{
\strong{Maintainer}: Fernando Cagua-Jahn \email{fernando.cagua@globalfishingwatch.org}

Authors:
\itemize{
  \item Fernando Cagua-Jahn \email{fernando.cagua@globalfishingwatch.org}
}

}
\keyword{internal}
